_backup_legacy/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
numpy>=1.26.0
pillow>=10.2.0
pyautogui>=0.9.54
selenium>=4.17.0
aiofiles>=23.2.1
langchain>=0.1.0
langgraph>=0.0.40
langchain-community>=0.0.20
langchain-core>=0.1.23
beautifulsoup4>=4.12.0
playwright>=1.41.0
httpx>=0.26.0
duckduckgo-search>=4.4.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
celery>=5.3.0
pymilvus>=2.3.0
weaviate-client>=3.24.0
pinecone-client>=3.0.0
prometheus-client>=0.19.0
python-json-logger>=2.0.0
cryptography>=41.0.0
pyjwt>=2.8.0
httpx[http2]>=0.26.0
tenacity>=8.2.0
structlog>=23.2.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-prometheus>=0.42b0
opentelemetry-exporter-jaeger>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-instrumentation-redis>=0.42b0
pydantic-extra-types>=2.4.0
